pyproject.toml
[tool.poetry]
name = "flex-usd"
version = "0.1.0"
description = "Token testbed and deployment for FlexUSD meant to be deployed on smartBCH"
authors = ["Sitt Guruvanich <sitt@coinflex.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7.10"
eth-brownie = "^1.16.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
bip-utils = "^1.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
norecursedirs = [".git", "build", "contracts", "interfaces", "outdated", "reports", "scripts"]
addopts = "--tb=short -s"
testpaths = ["tests"]